requirements.txt
# Jira to Rally Status Sync - Dependencies
jira>=3.4.1
pyral>=1.5.1
python-dotenv>=0.19.0
requests>=2.28.1

# GUI Application Dependencies  
# tkinter is included with Python by default
# Additional GUI enhancements (optional)
Pillow>=8.0.0  # For better image handling in GUI